First commit for ledger-parser
Ledger Parser is an attempt to create a backend javascript service
capable of processing ledger-cli files.

This first commit marks the beginning of this work and not much should
be expected right now.
All this module can currently do is read a ledger file and parse it into
JSON format. More features shall be added in the near feature, until the
app is actually capable of fulfilling my daily needs.