Add support for saving new entries
Ledger parser now has support for adding new entries to your accounting
file. Information on how to do this should be found in the readme file.

For ease of use with a RESTful API implementation the method for adding
new entries return a success/failure code (HTTP compliant) as well as a
friendly message, eg "Sucess!". As of this version, this cannot be
easily overriden yet, but will be made customisable in the near future.